day_04_music_genre_classifier/requirements.txt
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.24.0
scikit-learn>=1.4.0
joblib>=1.3.2
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.2
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0
python-dotenv>=1.0.0
loguru>=0.7.2
streamlit>=1.35.0
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.6
audioread>=3.0.1
